Généralise le nouveau composant chart
