+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/Indicateur/detail/IndicateurDetailChart.tsx
@@ -60,9 +60,7 @@ const IndicateurDetailChart = ({
       <IndicateurChartNew
         data={data}
         isLoading={isLoadingValeurs}
-        height={500}
         title={titre}
-        disableToolbox={!rempli}
       />
 
       {!!metadonnee && (
